second commit

+++ resources/views/siswa/index.blade.php
@@ -21,7 +21,9 @@
                 <th scope="col">Jenis Kelamin</th>
                 <th scope="col">Nama Kelas</th>
                 <th scope="col">Telepon</th>
+                @if (Auth::check())
                 <th scope="row">Action</th>
+                @endif
             </tr>
         </thead>
         <tbody>
@@ -33,19 +35,21 @@
                 <td>{{ $siswa->jenis_kelamin}}</td>
                 <td>{{ $siswa->kelas->nama_kelas}}</td>
                 <td>{{ !empty($siswa->telepon->nomor_telepon) ? $siswa->telepon->nomor_telepon : '-' }}</td>
-                <td>
+                @if (Auth::check())
+                    <td>
                     <div class="box-button">
                         {{ link_to('siswa/'.$siswa->id, 'Detail', ['class'=> 'btn btn-success btn-sm' ]) }}
-                    </div>
-                    <div class="box-button">
-                        {{ link_to('siswa/'.$siswa->id.'/edit', 'Edit', ['class'=> 'btn btn-warning btn-sm' ]) }}
-                    </div>
-                    <div class="box-button">
-                          {!! Form::open(['method' => 'DELETE', 'action'=> ['SiswaController@destroy',$siswa->id] ]) !!}
-                          {!! Form::submit('Delete',['class' => 'btn btn-danger btn-sm']) !!}                              
-                          {!! Form::close() !!}                              
-                    </div>
-                </td>
+                        </div>
+                        <div class="box-button">
+                            {{ link_to('siswa/'.$siswa->id.'/edit', 'Edit', ['class'=> 'btn btn-warning btn-sm' ]) }}
+                        </div>
+                        <div class="box-button">
+                            {!! Form::open(['method' => 'DELETE', 'action'=> ['SiswaController@destroy',$siswa->id] ]) !!}
+                            {!! Form::submit('Delete',['class' => 'btn btn-danger btn-sm']) !!}                              
+                            {!! Form::close() !!}                              
+                        </div>
+                    </td>
+                @endif 
             </tr>
             @endforeach 
         </tbody>
@@ -64,10 +68,11 @@
                 {{ $siswa_list->links() }}
             </div>
     </div>
-
-    <div class="bottom-nav">
-        <a href="siswa/create" class="btn btn-primary"> Tambah Siswa</a>
-    </div>
+    @if (Auth::check())
+       <div class="bottom-nav">
+         <a href="siswa/create" class="btn btn-primary"> Tambah Siswa</a>
+       </div>
+    @endif
 </div>
 @endsection
